# contracts/openapi/orders.yaml
openapi: 3.1.0
info:
  title: Waypoint Order Service
  version: 0.2.0
servers:
  - url: /api/v1
paths:
  /orders:
    get:
      operationId: listOrders
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: brand, in: query, schema: { type: string } }
        - { name: outlet_id, in: query, schema: { type: string } }
        - { name: requested_delivery_date, in: query, schema: { type: string, format: date } }
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Order" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createOrder
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrder" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  order: { $ref: "#/components/schemas/Order" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
  /orders/{id}:
    get:
      operationId: getOrder
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                type: object
                properties:
                  order: { $ref: "#/components/schemas/Order" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /orders/receipts/pending:
    get:
      operationId: listPendingReceipts
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Pending receipt confirmations for authorized outlet or dispatcher view }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /orders/receipt-issues:
    get:
      operationId: listReceiptIssues
      security: [{ bearerAuth: [] }]
      description: Dispatcher-only read-only receipt issue view.
      responses:
        "200": { description: Reported receipt issues }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /orders/{id}/tracking:
    get:
      operationId: getOrderTracking
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/ID" }]
      responses:
        "200": { description: Aggregated order, planning, delivery, and receipt state }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /orders/{id}/receipt:
    get:
      operationId: getOrderReceipt
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/ID" }]
      responses:
        "200": { description: Receipt and discrepancy records }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /orders/{id}/receipt/confirm:
    post:
      operationId: confirmOrderReceipt
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/ID" }, { name: Idempotency-Key, in: header, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [receivedUnits]
              properties:
                receivedUnits: { type: integer, minimum: 1 }
                issue: { $ref: "#/components/schemas/ReceiptIssueRequest" }
      responses:
        "201": { description: Receipt confirmed }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { description: Delivery has no receivable terminal outcome or receipt conflicts }
  /orders/{id}/receipt/issues:
    post:
      operationId: reportReceiptIssue
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/ID" }, { name: Idempotency-Key, in: header, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReceiptIssueRequest" }
      responses:
        "201": { description: Receipt issue recorded }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { description: No receipt or idempotency conflict }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: string }
  schemas:
    CreateOrder:
      type: object
      required: [requestedDeliveryDate, orderUnits, orderWeightKg, orderVolumeM3, temperatureRequirement]
      properties:
        requestedDeliveryDate: { type: string, format: date }
        orderUnits: { type: integer, exclusiveMinimum: 0 }
        orderWeightKg: { type: number, exclusiveMinimum: 0 }
        orderVolumeM3: { type: number, exclusiveMinimum: 0 }
        temperatureRequirement: { type: string, enum: [ambient, chilled] }
    Order:
      type: object
      properties:
        id: { type: string }
        orderRef: { type: string }
        outletId: { type: string }
        brand: { type: string }
        requestedDeliveryDate: { type: string }
        orderUnits: { type: integer }
        orderWeightKg: { type: number }
        orderVolumeM3: { type: number }
        temperatureRequirement: { type: string }
        status: { type: string }
        createdBy: { type: string }
    ReceiptIssueRequest:
      type: object
      required: [issueType, affectedUnits, idempotencyKey]
      properties:
        issueType: { type: string, enum: [MISSING, DAMAGED, QUANTITY_MISMATCH, OTHER] }
        affectedUnits: { type: integer, minimum: 0 }
        note: { type: string, maxLength: 1000 }
        idempotencyKey: { type: string }
  responses:
    BadRequest:
      description: Invalid request
    Unauthorized:
      description: Unauthenticated
    Forbidden:
      description: Not authorized
    NotFound:
      description: Missing
    Internal:
      description: Unexpected error